Moved to src-layout
Basically working from https://setuptools.pypa.io/en/latest/userguide/package_discovery.html.

I think this can still do with some improvement; `from filmcalendar.filmcalendar import FilmCalendar` seems a bit overwrought. But let's capture a working version while it's working.

Note: add `seattle` to the list of things which needs to be generalized in #6.
