+++ src/filmcalendar/seattle/siff.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta
 import requests
 from bs4 import BeautifulSoup
 
-from . import filmcalendar
+from filmcalendar import filmcalendar
 
 
 class FilmCalendarSIFF(filmcalendar.FilmCalendar):
