+++ src/filmcalendar/seattle/thebeacon.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import requests
 from bs4 import BeautifulSoup
 
-from . import filmcalendar
+from filmcalendar import filmcalendar
 
 
 class FilmCalendarTheBeacon(filmcalendar.FilmCalendar):
